requirements.txt
annotated-types==0.7.0
anyio==4.9.0
apscheduler==3.10.4
certifi==2025.7.14
dnspython==2.7.0
exchange_calendars==4.10.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
korean-lunar-calendar==0.3.1
numpy==2.3.1
pandas==2.3.1
pandas_market_calendars==5.1.1
pydantic==2.11.7
pydantic_core==2.33.2
pyluach==2.2.0
pymongo==4.13.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
tastytrade>=10.3.0
toolz==1.0.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
websockets==15.0.1
python-dotenv==1.1.1
